fix(autocompleter): Patch autocompleter to avoid scope confusion in iframe
While testing with Cypress, I've experienced some scope confusions because
of the sandboxing environment using iframe.
It was due to call of global values like window and window.document.
Also, set the doc document variable into the attach function to avoid
scope confusion when the 'input' node is not mounted to the page.